Move IND and ROD to rank.py
